make only webpack in a separate process
