+++ lib/core/engine.js
@@ -8,10 +8,7 @@ const CodeGenerator = require('../contracts/code_generator.js');
 const ServicesMonitor = require('./services_monitor.js');
 const Watch = require('../pipeline/watch.js');
 const LibraryManager = require('../versions/library_manager.js');
-const utils = require('../utils/utils');
-const constants = require('../constants');
-const async = require('async');
-const fs = require('../core/fs.js');
+let Pipeline = require('../pipeline/pipeline.js');
 
 class Engine {
   constructor(options) {
@@ -134,81 +131,27 @@ class Engine {
   }
 
   pipelineService(_options) {
-    let self = this;
-    self.events.emit("status", "Building Assets");
-    const pipelineProcess = require('child_process').fork(utils.joinPath(__dirname, '../pipeline/pipeline.js'));
-
-    async.waterfall([
-      function getWeb3Version(next) {
-        self.events.request("version:get:web3", function(web3Version) {
-          next(null, web3Version);
-        });
-      },
-      function getWeb3Location(web3Version, next) {
-        if (web3Version === "1.0.0-beta") {
-          return next(null, utils.joinPath(fs.embarkPath("js/web3-1.0.min.js")));
-        } else {
-          self.events.request("version:getPackageLocation", "web3", web3Version, function(err, location) {
-            return next(err, fs.dappPath(location));
-          });
-        }
-      },
-      function getProviderCode(web3Location, next) {
-        self.events.once('code-generator-ready', function () {
-          self.events.request('provider-code', function(providerCode) {
-            next(null, web3Location, providerCode);
-          });
-        });
-      }
-    ], (err, web3Location, providerCode) => {
-      if (err) {
-        return self.logger.error(err);
-      }
-
-      // Setup Pipeline
-      pipelineProcess.send({action: constants.pipeline.init, options: {
-          buildDir: self.config.buildDir,
-          contractsFiles: self.config.contractsFiles,
-          assetFiles: self.config.assetFiles,
-          events: self.events,
-          pipelinePlugins: self.plugins.getPluginsFor('pipeline'),
-          pluginImports: self.plugins.getPluginsProperty('imports', 'imports'),
-          web3Location,
-          providerCode
-        }});
-
-      pipelineProcess.on('message', function (msg) {
-        if (msg.result === constants.pipeline.built) {
+    const self = this;
+    this.events.emit("status", "Building Assets");
+    const pipeline = new Pipeline({
+      buildDir: this.config.buildDir,
+      contractsFiles: this.config.contractsFiles,
+      assetFiles: this.config.assetFiles,
+      events: this.events,
+      logger: this.logger,
+      plugins: this.plugins
+    });
+    this.events.on('code-generator-ready', function () {
+      self.events.request('code', function (abi, contractsJSON) {
+        self.currentAbi = abi;
+        self.contractsJSON = contractsJSON;
+        pipeline.build(abi, contractsJSON, null, function() {
           if (self.watch) {
             self.watch.restart(); // Necessary because changing a file while it is writing can stop it from being watched
           }
-          if (msg.error) {
-            return self.logger.error(msg.error);
-          }
-          return self.events.emit('outputDone');
-        }
-
-        if (msg.result === constants.pipeline.log) {
-          if (self.logger[msg.type]) {
-            return self.logger[msg.type](self.normalizeInput(msg.message));
-          }
-          self.logger.debug(self.normalizeInput(msg.message));
-        }
-      });
-
-      function build() {
-        self.events.request('code', function (abi, contractsJSON) {
-          self.currentAbi = abi;
-          self.contractsJSON = contractsJSON;
-          pipelineProcess.send({action: constants.pipeline.build, abi, contractsJSON, path: null});
+          self.events.emit('outputDone');
         });
-      }
-
-      self.events.on('code-generator-ready', function () {
-        build();
       });
-
-      build();
     });
   }
