+++ lib/pipeline/pipeline.js
@@ -1,7 +1,7 @@
 const fs = require('../core/fs.js');
 const async = require('async');
+const child_process = require('child_process');
 const utils = require('../utils/utils.js');
-const webpack = require("webpack");
 const constants = require('../constants');
 const File = require('../core/file');
 
@@ -10,45 +10,15 @@ require("babel-preset-es2015");
 require("babel-preset-es2016");
 require("babel-preset-es2017");
 
-let pipeline;
-
-// Override process.chdir so that we have a partial-implementation PWD for Windows
-const realChdir = process.chdir;
-process.chdir = (...args) => {
-  if (!process.env.PWD) {
-    process.env.PWD = process.cwd();
-  }
-  realChdir(...args);
-};
-
 class Pipeline {
 
   constructor(options) {
     this.buildDir = options.buildDir;
     this.contractsFiles = options.contractsFiles;
     this.assetFiles = options.assetFiles;
-    this.pipelinePlugins = options.pipelinePlugins;
-    this.pluginImports = options.pluginImports;
-    this.web3Location = options.web3Location;
-    this.providerCode = options.providerCode;
-
-    this.interceptLogs();
-  }
-
-  interceptLogs() {
-    const context = {};
-    context.console = console;
-
-    context.console.log = this.log.bind(this, 'log');
-    context.console.warn = this.log.bind(this, 'warn');
-    context.console.info = this.log.bind(this, 'info');
-    context.console.debug = this.log.bind(this, 'debug');
-    context.console.trace = this.log.bind(this, 'trace');
-    context.console.dir = this.log.bind(this, 'dir');
-  }
-
-  log(type, ...messages) {
-    process.send({result: constants.pipeline.log, message: messages, type});
+    this.events = options.events;
+    this.logger = options.logger;
+    this.plugins = options.plugins;
   }
 
   build(abi, contractsJSON, path, callback) {
@@ -66,7 +36,7 @@ class Pipeline {
         importsList["Embark/EmbarkJS"] = fs.dappPath(".embark", 'embark.js');
         importsList["Embark/web3"] = fs.dappPath(".embark", 'web3_instance.js');
 
-        self.pluginImports.forEach(function (importObject) {
+        self.plugins.getPluginsProperty('imports', 'imports').forEach(function (importObject) {
           let [importName, importLocation] = importObject;
           importsList[importName] = importLocation;
         });
@@ -89,7 +59,7 @@ class Pipeline {
             async.mapLimit(files, 1,
               function (file, fileCb) {
                 file = new File(file); // Re-instantiate a File as through the process, we lose its prototype
-                console.trace("reading " + file.filename);
+                self.logger.trace("reading " + file.filename);
 
                 if (file.filename.indexOf('.js') < 0) {
                   return file.content(function (fileContent) {
@@ -98,33 +68,34 @@ class Pipeline {
                 }
 
                 // JS files
-                let realCwd;
-
                 async.waterfall([
-
-                  function findImports(next) {
-                    self.webpackRun(file.filename, {}, false, importsList, false, next);
-                  },
-
-                  function changeCwd(next) {
-                    realCwd = utils.pwd();
-                    process.chdir(fs.embarkPath(''));
-                    next();
-                  },
-
                   function runWebpack(next) {
-                    self.webpackRun(file.filename, {}, true, importsList, true, next);
-                  },
+                    const webpackProcess = child_process.fork(utils.joinPath(__dirname, 'webpackProcess.js'));
+                    webpackProcess.send({action: constants.pipeline.init, options: {}});
+                    webpackProcess.send({action: constants.pipeline.build, file, importsList});
 
-                  function changeCwdBack(next) {
-                    process.chdir(realCwd);
-                    next();
+                    webpackProcess.on('message', function (msg) {
+                      if (msg.result === constants.pipeline.built) {
+                        webpackProcess.disconnect();
+                        if (msg.error) {
+                          return next(msg.error);
+                        }
+                        return next();
+                      }
+
+                      if (msg.result === constants.pipeline.log) {
+                        if (self.logger[msg.type]) {
+                          return self.logger[msg.type](self.normalizeInput(msg.message));
+                        }
+                        self.logger.debug(self.normalizeInput(msg.message));
+                      }
+                    });
                   },
 
                   function checkFile(next) {
                     fs.access('./.embark/' + file.filename, (err) => {
                       if (err) {
-                        console.error("couldn't find file: " + file.filename);
+                        self.logger.error("couldn't find file: " + file.filename);
                         return next("couldn't find file: " + file.filename);
                       }
                       next();
@@ -146,8 +117,7 @@ class Pipeline {
 
                 ], function (err, contentFile) {
                   if (err) {
-                    process.chdir(realCwd);
-                    console.error(err);
+                    self.logger.error(err);
                     return fileCb(err);
                   }
 
@@ -156,10 +126,10 @@ class Pipeline {
               },
               function (err, contentFiles) {
                 if (err) {
-                  console.error('errors found while generating ' + targetFile);
+                  self.logger.error('errors found while generating ' + targetFile);
                 }
                 let dir = targetFile.split('/').slice(0, -1).join('/');
-                console.trace("creating dir " + self.buildDir + dir);
+                self.logger.trace("creating dir " + self.buildDir + dir);
                 fs.mkdirpSync(self.buildDir + dir);
 
                 // if it's a directory
@@ -172,7 +142,7 @@ class Pipeline {
 
                   async.each(contentFiles, function (file, mapCb) {
                     let filename = file.filename.replace(file.basedir + '/', '');
-                    console.info("writing file " + (self.buildDir + targetDir + filename).bold.dim);
+                    self.logger.info("writing file " + (self.buildDir + targetDir + filename).bold.dim);
 
                     fs.copy(file.path, self.buildDir + targetDir + filename, {overwrite: true}, mapCb);
                   }, cb);
@@ -186,7 +156,7 @@ class Pipeline {
                   return file.content;
                 }).join("\n");
 
-                console.info("writing file " + (self.buildDir + targetFile).bold.dim);
+                self.logger.info("writing file " + (self.buildDir + targetFile).bold.dim);
                 fs.writeFile(self.buildDir + targetFile, content, cb);
               }
             );
@@ -198,17 +168,12 @@ class Pipeline {
 
   runPlugins(file, fileContent, fileCb) {
     const self = this;
-    if (self.pipelinePlugins.length <= 0) {
-      return fileCb(null, {
-        content: fileContent,
-        filename: file.filename,
-        path: file.path,
-        basedir: file.basedir,
-        modified: true
-      });
+    let pipelinePlugins = self.plugins.getPluginsFor('pipeline');
+    if (pipelinePlugins.length <= 0) {
+      return fileCb(null, {content: fileContent, filename: file.filename, path: file.path, basedir: file.basedir, modified: true});
     }
-    async.eachSeries(self.pipelinePlugins,
-      function (plugin, pluginCB) {
+    async.eachSeries(pipelinePlugins,
+      function(plugin, pluginCB) {
         if (file.options && file.options.skipPipeline) {
           return pluginCB();
         }
@@ -219,83 +184,11 @@ class Pipeline {
       },
       function (err) {
         if (err) {
-          console.error(err.message);
+          self.logger.error(err.message);
         }
-        return fileCb(null, {
-          content: fileContent,
-          filename: file.filename,
-          path: file.path,
-          basedir: file.basedir,
-          modified: true
-        });
+        return fileCb(null, {content: fileContent, filename: file.filename, path: file.path, basedir: file.basedir, modified: true});
       }
     );
-  }
-
-  webpackRun(filename, options, includeModules, importsList, detectErrors, callback) {
-    let defaultOptions = {
-      entry: fs.dappPath(filename),
-      output: {
-        libraryTarget: 'umd',
-        path: fs.dappPath('.embark'),
-        filename: filename
-      },
-      resolve: {
-        alias: importsList,
-        modules: [
-          fs.embarkPath('node_modules'),
-          fs.dappPath('node_modules')
-        ]
-      },
-      externals: function (context, request, callback) {
-        callback();
-      }
-    };
-
-    let webpackOptions = utils.recursiveMerge(defaultOptions, options);
-
-    if (includeModules) {
-      webpackOptions.module = {
-        rules: [
-          {
-            test: /\.css$/,
-            use: [{loader: "style-loader"}, {loader: "css-loader"}]
-          },
-          {
-            test: /\.scss$/,
-            use: [{loader: "style-loader"}, {loader: "css-loader"}]
-          },
-          {
-            test: /\.(png|woff|woff2|eot|ttf|svg)$/,
-            loader: 'url-loader?limit=100000'
-          },
-          {
-            test: /\.js$/,
-            loader: "babel-loader",
-            exclude: /(node_modules|bower_components)/,
-            options: {
-              presets: ['babel-preset-es2016', 'babel-preset-es2017', 'babel-preset-react'].map(require.resolve),
-              plugins: ["babel-plugin-webpack-aliases"].map(require.resolve),
-              compact: false
-            }
-          }
-        ]
-      };
-    }
-
-    webpack(webpackOptions).run((err, stats) => {
-      if (err) {
-        console.error(err);
-      }
-      if (!detectErrors) {
-        return callback();
-      }
-
-      if (stats.hasErrors()) {
-        return callback(stats.toJson().errors.join("\n"));
-      }
-      callback();
-    });
   }
 
   buildContracts(contractsJSON, callback) {
@@ -337,21 +230,33 @@ class Pipeline {
     let code = "";
 
     async.waterfall([
-      function getImports(next) {
-        self.web3Location = self.web3Location.replace(/\\/g, '/'); // Import paths must always have forward slashes
-        code += "\nimport Web3 from '" + self.web3Location + "';\n";
+      function getWeb3Location(next) {
+        self.events.request("version:get:web3", function(web3Version) {
+          if (web3Version === "1.0.0-beta") {
+            return next(null, utils.joinPath(fs.embarkPath("js/web3-1.0.min.js")));
+          } else {
+            self.events.request("version:getPackageLocation", "web3", web3Version, function(err, location) {
+              return next(null, fs.dappPath(location));
+            });
+          }
+        });
+      },
+      function getImports(web3Location, next) {
+        web3Location = web3Location.replace(/\\/g, '/'); // Import paths must always have forward slashes
+        code += "\nimport Web3 from '" + web3Location + "';\n";
 
         code += "\n if (typeof web3 !== 'undefined') {";
         code += "\n } else {";
         code += "\n var web3 = new Web3();\n";
         code += "\n }";
 
-        code += self.providerCode;
-        code += "\nglobal.__embarkContext = __mainContext.__loadManagerInstance;\n";
-        code += "\nwindow.web3 = web3;\n";
-        code += "\nexport default web3;\n";
-
-        next();
+        self.events.request('provider-code', function(providerCode)  {
+          code += providerCode;
+          code += "\nglobal.__embarkContext = __mainContext.__loadManagerInstance;\n";
+          code += "\nwindow.web3 = web3;\n";
+          code += "\nexport default web3;\n";
+          next();
+        });
       },
       function makeDirectory(next) {
         fs.mkdirp(fs.dappPath(".embark"), (err, _result) => {
@@ -365,19 +270,4 @@ class Pipeline {
   }
 }
 
-process.on('message', (msg) => {
-  if (msg.action === constants.pipeline.init) {
-    pipeline = new Pipeline(msg.options);
-    return process.send({result: constants.pipeline.initiated});
-  }
-
-  if (msg.action === constants.pipeline.build) {
-    return pipeline.build(msg.abi, msg.contractsJSON, msg.path, (err) => {
-      process.send({result: constants.pipeline.built, error: err});
-    });
-  }
-});
-
-process.on('exit', () => {
-  process.exit(0);
-});
+module.exports = Pipeline;
